Drop hardcoded path from mktemp in job scripts
The `mktemp` command creates a temporary file or directory.  This change
removes the hard-coded `/tmp` path from the two places where `mktemp` is
executed in a job script.

The reason is, in UNIX operating systems (including Linux), there is a
standard[1] environment variable, `TMPDIR`, which system administrators
can tell programs to store data.  On SCG, compute jobs have `TMPDIR` set
to a path that is automatically cleaned up when the job completes.

When you set a hard-coded path in `mktemp`, the command ignores the
`TMPDIR`.  Instead, it is better to either not provide a template at
all, or to provide a template _and_ use the `--tmpdir` option (which
makes your path relative to `TMPDIR`).

[1] IEEE.  Portable Operating System Interface (POSIX(TM)) Base Specifications.
    Issue 7.  Chapter 8.  DOI 10.1109/IEEESTD.2018.8277153.
    Free link at  https://pubs.opengroup.org/onlinepubs/9699919799/